+++ changelog.md
@@ -1,3 +1,5 @@
 **Changes:**
-- Added French, Japanese, Portuguese and Russian language
-- Fixed Fabric Catalogue item icon
+- Added two new music discs from Kevin MacLeod
+  - "Guzheng City" can be found in the village toolsmith chest
+  - "Night Cave" can be found in the stronghold library chest
+- Updated dependencies